refactor : errorMessages.js로 에러메세지를 관리
